fixed issue of image path
