+++ apps/web-giddh/src/app/shared/carousel/carousel.component.scss
@@ -17,8 +17,6 @@
 
     .carousel-content {
         width: min(var(--carousel-content-width), 80svw);
-        max-height: 80svh;
-        min-height: 70svh;
     }
 
     .slider-icon {
